Move helper classes out of YAML namespace
