+++ resharper/resharper-unity/src/CSharp/Daemon/Stages/Analysis/LoadSceneAnalyzer.cs
@@ -1,6 +1,7 @@
 using JetBrains.ProjectModel;
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Errors;
+using JetBrains.ReSharper.Plugins.Unity.Utils;
 using JetBrains.ReSharper.Plugins.Unity.Yaml;
 using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
@@ -112,7 +113,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
                 return null;
 
             var sceneName = constantValue;
-            if (sceneName.StartsWith("Assets/") && sceneName.EndsWith(UnityYamlFileExtensions.SceneFileExtensionWithDot))
+            if (sceneName.StartsWith("Assets/") && sceneName.EndsWith(UnityAssetFileExtensions.SceneFileExtensionWithDot))
                 sceneName = UnityProjectSettingsUtils.GetUnityScenePathRepresentation(sceneName);
 
             return sceneName;
