+++ resharper/resharper-unity/src/Core/Rider/Psi/Modules/RiderUnityExternalFilesIndexDisablingStrategy.cs
@@ -3,7 +3,6 @@ using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.Caches;
 using JetBrains.ReSharper.Plugins.Unity.Core.Psi.Modules;
 using JetBrains.ReSharper.Plugins.Unity.Rider.Protocol;
-using JetBrains.ReSharper.Plugins.Unity.Yaml;
 using JetBrains.ReSharper.Psi.Util;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Core.Rider.Psi.Modules
