+++ resharper/resharper-unity/src/Core/VisualStudio/Psi/Modules/ReSharperUnityExternalFilesIndexDisablingStrategy.cs
@@ -3,7 +3,6 @@ using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.Caches;
 using JetBrains.ReSharper.Plugins.Unity.Core.Psi.Modules;
-using JetBrains.ReSharper.Plugins.Unity.Yaml;
 using JetBrains.ReSharper.Psi.Util;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Core.VisualStudio.Psi.Modules
