+++ resharper/resharper-unity/src/Rider/UnityEditorFindUsageResultCreator.cs
@@ -12,6 +12,7 @@ using JetBrains.ProjectModel;
 using JetBrains.RdBackend.Common.Features.BackgroundTasks;
 using JetBrains.ReSharper.Plugins.Unity.Core.Psi.Modules;
 using JetBrains.ReSharper.Plugins.Unity.Rider.Protocol;
+using JetBrains.ReSharper.Plugins.Unity.Utils;
 using JetBrains.ReSharper.Plugins.Unity.Yaml;
 using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi;
 using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.DeferredCaches.AnimatorUsages;
@@ -125,7 +126,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider
                     isStateMachine ? AnimatorUsageType.StateMachine : AnimatorUsageType.State, needExpand,
                     pathFromAsset, fileName, extension);
             }
-            if (path.ExtensionWithDot.EndsWith(UnityYamlFileExtensions.AssetFileExtensionWithDot))
+            if (path.ExtensionWithDot.EndsWith(UnityAssetFileExtensions.AssetFileExtensionWithDot))
             {
                 return new AssetFindUsagesResult(needExpand, pathFromAsset, fileName, extension);
             }
