+++ resharper/resharper-unity/src/Yaml/Psi/Caches/TagManagerAssetHandler.cs
@@ -1,4 +1,5 @@
 using JetBrains.ProjectModel;
+using JetBrains.ReSharper.Plugins.Unity.Utils;
 using JetBrains.ReSharper.Plugins.Yaml.Psi;
 using JetBrains.ReSharper.Plugins.Yaml.Psi.Tree;
 using JetBrains.ReSharper.Psi;
