+++ resharper/resharper-unity/src/Yaml/Psi/UnityYamlPsiSourceFileExtensions.cs
@@ -1,3 +1,4 @@
+using JetBrains.ReSharper.Plugins.Unity.Utils;
 using JetBrains.ReSharper.Psi;
 
 namespace JetBrains.ReSharper.Plugins.Unity.Yaml.Psi
